+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/CheckReturnValue.java
@@ -1,0 +1,7 @@
+package nl.jqno.equalsverifier;
+
+import java.lang.annotation.*;
+
+@Target({ ElementType.CONSTRUCTOR, ElementType.METHOD, ElementType.PACKAGE, ElementType.TYPE })
+@Retention(RetentionPolicy.CLASS)
+public @interface CheckReturnValue {}