+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/ConfiguredEqualsVerifier.java
@@ -48,6 +48,7 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
      *
      * @return a copy of the configuration.
      */
+    @CheckReturnValue
     public ConfiguredEqualsVerifier copy() {
         return new ConfiguredEqualsVerifier(EnumSet.copyOf(warningsToSuppress),
                 factoryCache.copy(),
@@ -57,6 +58,7 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
 
     /** {@inheritDoc} */
     @Override
+    @CheckReturnValue
     public ConfiguredEqualsVerifier suppress(Warning... warnings) {
         Collections.addAll(warningsToSuppress, warnings);
         return this;
@@ -64,6 +66,7 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
 
     /** {@inheritDoc} */
     @Override
+    @CheckReturnValue
     public <S> ConfiguredEqualsVerifier withPrefabValues(Class<S> otherType, S red, S blue) {
         PrefabValuesApi.addPrefabValues(factoryCache, objenesis, otherType, red, blue);
         return this;
@@ -71,6 +74,7 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
 
     /** {@inheritDoc} */
     @Override
+    @CheckReturnValue
     public <S> ConfiguredEqualsVerifier withGenericPrefabValues(Class<S> otherType, Func1<?, S> factory) {
         PrefabValuesApi.addGenericPrefabValues(factoryCache, otherType, factory);
         return this;
@@ -78,6 +82,7 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
 
     /** {@inheritDoc} */
     @Override
+    @CheckReturnValue
     public <S> ConfiguredEqualsVerifier withGenericPrefabValues(Class<S> otherType, Func2<?, ?, S> factory) {
         PrefabValuesApi.addGenericPrefabValues(factoryCache, otherType, factory);
         return this;
@@ -85,12 +90,14 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
 
     /** {@inheritDoc} */
     @Override
+    @CheckReturnValue
     public ConfiguredEqualsVerifier usingGetClass() {
         usingGetClass = true;
         return this;
     }
 
     @Override
+    @CheckReturnValue
     public ConfiguredEqualsVerifier withFieldnameToGetterConverter(Function<String, String> converter) {
         this.fieldnameToGetter = converter;
         return this;
@@ -98,11 +105,12 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
 
     /**
      * {@inheritDoc}
-     * 
+     *
      * @deprecated No longer needed; this happens automatically.
      */
     @Deprecated
     @Override
+    @CheckReturnValue
     public ConfiguredEqualsVerifier withResetCaches() {
         return this;
     }
@@ -114,6 +122,7 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
      * @param type The class for which the {@code equals} method should be tested.
      * @return A fluent API for EqualsVerifier.
      */
+    @CheckReturnValue
     public <T> SingleTypeEqualsVerifierApi<T> forClass(Class<T> type) {
         return new SingleTypeEqualsVerifierApi<>(type,
                 EnumSet.copyOf(warningsToSuppress),
@@ -129,6 +138,7 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
      * @param classes An iterable containing the classes for which {@code equals} method should be tested.
      * @return A fluent API for EqualsVerifier.
      */
+    @CheckReturnValue
     public MultipleTypeEqualsVerifierApi forClasses(Iterable<Class<?>> classes) {
         return new MultipleTypeEqualsVerifierApi(ListBuilders.fromIterable(classes), this);
     }
@@ -141,6 +151,7 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
      * @param more   More classes for which the {@code equals} method should be tested.
      * @return A fluent API for EqualsVerifier.
      */
+    @CheckReturnValue
     public MultipleTypeEqualsVerifierApi forClasses(Class<?> first, Class<?> second, Class<?>... more) {
         return new MultipleTypeEqualsVerifierApi(ListBuilders.buildListOfAtLeastTwo(first, second, more), this);
     }
@@ -155,6 +166,7 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
      * @param packageName A package for which each class's {@code equals} should be tested.
      * @return A fluent API for EqualsVerifier.
      */
+    @CheckReturnValue
     public MultipleTypeEqualsVerifierApi forPackage(String packageName) {
         return forPackage(packageName, false);
     }
@@ -170,6 +182,7 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
      * @param scanRecursively true to scan all sub-packages
      * @return A fluent API for EqualsVerifier.
      */
+    @CheckReturnValue
     public MultipleTypeEqualsVerifierApi forPackage(String packageName, boolean scanRecursively) {
         List<Class<?>> classes = PackageScanner.getClassesIn(packageName, null, scanRecursively);
         Validations.validatePackageContainsClasses(packageName, classes);
@@ -190,6 +203,7 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
      * @param mustExtend  if not null, returns only classes that extend or implement this class.
      * @return A fluent API for EqualsVerifier.
      */
+    @CheckReturnValue
     public MultipleTypeEqualsVerifierApi forPackage(String packageName, Class<?> mustExtend) {
         List<Class<?>> classes = PackageScanner.getClassesIn(packageName, mustExtend, true);
         Validations.validatePackageContainsClasses(packageName, classes);